% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic Function.R
\name{mean_per_year}
\alias{mean_per_year}
\title{Mean number of passengers in selected months per year.}
\usage{
mean_per_year(dataset, ...)
}
\arguments{
\item{dataset}{A data set.}
}
\value{
The mean number of airline passengers per year in the uk.
}
\description{
This is a generic function which takes a data set as its argument.
}
\details{
Each column in the data set should represent one specific year.

If the argument does not have the class of data frame or numeric
the function will return the class of the function.

\code{mean_per_year} returns the mean of each column
or vector present in the data set if has the class of data frame or numeric.
}
\examples{
year_2015 <- c(34, 67, 71)
mean_per_year(year_2015)

number_of_airpassengers_in_uk is a data frame
mean_per_year(number_of_airpassengers_in_uk)
}
