% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_airpassengers_in_uk.R
\docType{data}
\name{number_of_airpassengers_in_uk}
\alias{number_of_airpassengers_in_uk}
\title{Number of airpassengers in the uk between june and august in 2016 - 2018}
\format{A data frame with 3 rows representing the months
and 3 columns representing the years:
\describe{
  \item{month}{month, e.g. june}
  \item{year}{number of airpassengers, in thousands}
}}
\usage{
number_of_airpassengers_in_uk
}
\description{
A dataset in the form of a data frame containing the number of airpassengers
in the uk during the months of june, july and august within the period of 2016 to 2018.
}
\keyword{datasets}
